checkout method
