+++ src/api/cart/backend-cart.js
@@ -8,31 +8,7 @@
 
 import _ from 'lodash';
 import * as endpoints from '../../endpoints';
-
-// reduce SKU list
-function collectLineItems(skus) {
-  const uniqueSkus = Object.create(null);
-  return _.reduce(skus, (result, lineItem) => {
-    const sku = lineItem.sku;
-
-    if (sku in uniqueSkus) {
-      const qty = result[uniqueSkus[sku]].quantity += lineItem.quantity;
-      result[uniqueSkus[sku]].totalPrice = lineItem.price * qty;
-    } else {
-      uniqueSkus[sku] = result.length;
-      result.push(lineItem);
-    }
-
-    return result;
-  }, []);
-}
-
-function normalizeResponse(payload) {
-  if (payload.lineItems) {
-    payload.lineItems.skus = collectLineItems(payload.lineItems.skus);
-  }
-  return payload;
-}
+import { normalizeResponse } from './common';
 
 export default class BackendCart {
   constructor(api) {
